[#66550694] Solving "permissions" for imageFileModel
--HG--
branch : imageUploadAndGalleryV2

+++ app/protected/modules/zurmo/controllers/ImageModelController.php
@@ -100,6 +100,10 @@
         public function actionDelete($id)
         {
             $imageFileModel = ImageFileModel::getById((int)$id);
+            if (!$imageFileModel->canDelete())
+            {
+                throw new NotSupportedException();
+            }
             $imageFileModel->delete();
         }
